small debug added

+++ ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: tcohen <tcohen@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/27 18:18:46 by tcohen            #+#    #+#             */
-/*   Updated: 2025/01/30 16:40:27 by tcohen           ###   ########.fr       */
+/*   Updated: 2025/02/04 15:27:59 by tcohen           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,8 @@ int main(int argc, char **argv)
     }
     str = argv[1];
     size_t len = str.length();
-    for (int i = 0; i < len; i++)
-        str[i] = std::toupper(str[i]);
+    for (size_t i = 0; i < len; i++)
+        	str[i] = std::toupper(str[i]);
     std::cout << str << std::endl;
     return (0);
 }
